Commit #10
Adjusted the motion blur effects of beer, then created wine, and
adjusted its blur effect, and finally repeated with hard Liqour. Then
adjusted the stills. Toilet wine, is the second fastest, produces the
least, but the safest. Beer bucket, is slowest, produces the most, and
has a chance of expolding. (small) Liquor still is the fastest, produces
the middle amount of liquor, but the highest chance of exploding. I
still have to work on Large Liquor Still. And have them produce liquor.
Til then the new drinks are added, to the buy menu, with testing level
prices.